requirements.txt
streamlit==1.35.0
yfinance==0.2.38
pandas==2.2.3
numpy==1.26.4
plotly==5.21.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.2.2
python-dotenv==1.0.1
openai==1.30.1
pytz==2024.1
serpapi==2.0.0